Clipyt/requirements.txt
opencv-python
tk
reportlab
yt-dlp
scikit-image
pytesseract
Pillow
